refreshed code /w Kevin
