--- backend/Dockerfile
@@ -1,18 +1,0 @@
-FROM node:carbon
-
-# Create app directory
-WORKDIR /usr/src/app
-
-# Install app dependencies
-# A wildcard is used to ensure both package.json AND package-lock.json are copied
-# where available (npm@5+)
-COPY package*.json ./
-
-RUN npm install
-
-# Bundle app source
-COPY . .
-
-EXPOSE 8080
-
-CMD [ "npm", "run", "serve" ]
