--- backend/config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    'Secret': 'aMqxxrfxrXs239SDRU6HF90eWfX8gaeL',
-    'MongoUrl': 'mongodb://admin:kevinrick@miniproject.rickvanfessem.nl:29017/miniproject?authSource=admin'
-}
