--- backend/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "nodejsapi",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon bin/www",
-    "serve": "node bin/www",
-    "build": "docker build -t miniprojectbackend .",
-    "tag": "docker tag miniprojectbackend rickvf/iotapi",
-    "push": "docker push rickvf/iotapi",
-    "publish": "npm run build && npm run tag && npm run push"
-  },
-  "dependencies": {
-    "bcrypt": "^1.0.3",
-    "body-parser": "~1.18.2",
-    "cookie-parser": "~1.4.3",
-    "cors": "^2.8.4",
-    "debug": "~2.6.9",
-    "express": "~4.15.5",
-    "jade": "~1.11.0",
-    "jsonwebtoken": "^8.2.0",
-    "moment": "^2.21.0",
-    "mongoose": "^5.0.7",
-    "morgan": "^1.9.0",
-    "nodemon": "^1.17.1",
-    "request": "^2.83.0",
-    "request-promise-native": "^1.0.5",
-    "serve-favicon": "^2.4.5"
-  }
-}
