+++ config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    'Secret': 'aMqxxrfxrXs239SDRU6HF90eWfX8gaeL',
+}
