+++ package.json
@@ -4,7 +4,7 @@
   "description": "The backend for the SCOUT Smartgreenhouse project",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js"
+    "start": "node bin/www"
   },
   "repository": {
     "type": "git",
@@ -19,6 +19,18 @@
   "dependencies": {
     "express": "^4.16.3",
     "jsonwebtoken": "^8.2.1",
-    "mongoose": "^5.0.15"
+    "mongoose": "^5.0.15",
+    "bcrypt": "^1.0.3",
+    "body-parser": "~1.18.2",
+    "cookie-parser": "~1.4.3",
+    "cors": "^2.8.4",
+    "debug": "~2.6.9",
+    "jade": "~1.11.0",
+    "moment": "^2.21.0",
+    "morgan": "^1.9.0",
+    "nodemon": "^1.17.1",
+    "request": "^2.83.0",
+    "request-promise-native": "^1.0.5",
+    "serve-favicon": "^2.4.5"
   }
 }
