+++ src/models/sensorType.js
@@ -1,0 +1,11 @@
+var mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+var schema = new Schema({
+    name: {type: String, required:true},
+},
+{
+    timestamps: true
+});
+
+module.exports = mongoose.model('SensorType', schema);
