+++ src/models/userAccount.js
@@ -1,0 +1,12 @@
+var mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+var schema = new Schema({
+    username: {type: String, required:true, unique: true},
+    password: {type: String, required:true},
+},
+{
+    timestamps: true
+});
+
+module.exports = mongoose.model('UserAccount', schema);
