+++ src/repository/repository.js
@@ -1,6 +1,4 @@
 const mongoose = require("mongoose");
-const config = require("../../config");
-
 class Repository {
   constructor() {}
